Single definition for task inputs/outputs
Define the inputs and outputs of the two task implementations in a
single file, pulled into both classes.

Fixes #171.
